requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-extensions==3.2.1
django-filter==21.1
djoser==2.1.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
flake8==6.0.0
Pillow==9.2.0
django-colorfield==0.10.1
djoser==2.1.0
drf-extra-fields==3.4.0
python-dotenv==0.21.0
psycopg2-binary==2.9.7
reportlab==4.0.6